+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - `Interval` field `time` renamed to `micros`.
 - `Interval` field `day` renamed to `days`.
 - `Interval` field `month` renamed to `months`.
+- The order of fields in `Interval` changed to `months`, `days`, `micros`.